Title: 쪼개어 배낭 채우기 구현, Time: 570ms, Memory: 24MB, Status: Wrong Answer - Codetree

+++ 251112/쪼개어 배낭 채우기 구현/implement-fractional-knapsack.java
@@ -35,7 +35,7 @@ public class Main {
 
         Arrays.sort(pairs,
                 Comparator.comparing(
-                        (Pair p) -> -(p.value / (double) p.wight)
+                        (Pair p) -> -(p.value /  p.wight)
                 )
         );
